refactor: split fixtures in separate files
