+++ tests/fixtures/1_starknet.py
@@ -1,0 +1,79 @@
+import os
+import logging
+import shutil
+from pathlib import Path
+from typing import AsyncGenerator
+
+import pandas as pd
+import pytest_asyncio
+from cairo_coverage import cairo_coverage
+from starkware.starknet.business_logic.execution.execute_entry_point import (
+    ExecuteEntryPoint,
+)
+from starkware.starknet.business_logic.state.state_api_objects import BlockInfo
+from starkware.starknet.testing.starknet import Starknet
+
+from tests.utils.utils import dump_coverage, dump_reports, timeit, traceit
+
+pd.set_option("display.max_rows", 500)
+pd.set_option("display.max_columns", 500)
+pd.set_option("display.width", 1000)
+
+logging.getLogger("asyncio").setLevel(logging.ERROR)
+logger = logging.getLogger()
+
+
+@pytest_asyncio.fixture(scope="session")
+async def starknet(worker_id, request) -> AsyncGenerator[Starknet, None]:
+    starknet = await Starknet.empty()
+    starknet.state.state.update_block_info(
+        BlockInfo.create_for_testing(block_number=1, block_timestamp=1)
+    )
+
+    starknet.deploy = traceit.trace_all(timeit(starknet.deploy))
+    starknet.declare = timeit(starknet.declare)
+    if request.config.getoption("trace_run"):
+        logger.info("trace-run option enabled")
+        ExecuteEntryPoint._run = traceit.trace_run(ExecuteEntryPoint._run)
+    else:
+        logger.info("trace-run option disabled")
+    output_dir = Path("coverage")
+    shutil.rmtree(output_dir, ignore_errors=True)
+
+    yield starknet
+
+    output_dir.mkdir(exist_ok=True, parents=True)
+    files = cairo_coverage.report_runs(
+        excluded_file={"site-packages", "tests"}
+    )
+    total_covered = []
+    for file in files:
+        if file.pct_covered < 80:
+            logger.warning(f"{file.name} only {file.pct_covered:.2f}% covered")
+        total_covered.append(file.pct_covered)
+    if files and (val := not sum(total_covered) / len(files)) >= 80:
+        logger.warning(f"Project is not covered enough {val:.2f})")
+
+    if worker_id == "master":
+        dump_reports(output_dir)
+        dump_coverage(output_dir, files)
+    else:
+        dump_reports(output_dir / worker_id)
+        dump_coverage(output_dir / worker_id, files)
+        if len(os.listdir(output_dir)) == int(os.environ["PYTEST_XDIST_WORKER_COUNT"]):
+            # This is the last teardown of the testsuite, merge the files
+            resources = pd.concat(
+                [pd.read_csv(f) for f in output_dir.glob("**/resources.csv")],
+            )
+            if not resources.empty:
+                resources.sort_values(["n_steps"], ascending=False).to_csv(
+                    output_dir / "resources.csv", index=False
+                )
+            times = pd.concat(
+                [pd.read_csv(f) for f in output_dir.glob("**/times.csv")],
+                ignore_index=True,
+            )
+            if not times.empty:
+                times.sort_values(["duration"], ascending=False).to_csv(
+                    output_dir / "times.csv", index=False
+                )
